By default, let read_image() automatically infer the dtype (np.uint8 or np.uint16) of the returned array.
It can handle both RGB8 (np.uint8) and gray16 (np.uint16) images.

The previous behavior (default dtype='np.uint8') produces invalid data for gray16 images, which was not useful.

PiperOrigin-RevId: 374272758

+++ mediapy_test.py
@@ -348,7 +348,7 @@ class MediapyTest(parameterized.TestCase):
     with tempfile.TemporaryDirectory() as directory_name:
       path = pathlib.Path(directory_name) / 'test.png'
       media.write_image(path, image)
-      new_image = media.read_image(path, dtype=dtype)
+      new_image = media.read_image(path)
       self.assert_all_equal(image.shape, new_image.shape)
       self.assertEqual(image.dtype, new_image.dtype)
       self.assert_all_equal(image, new_image)
